Add ignoring of incorrect exp and different key order
