+++ golikov.valentin/T2/local.h
@@ -1,0 +1,18 @@
+#ifndef LOCAL_H
+#define LOCAL_H
+
+//В этом файле подключаются все используемые библиотеки
+
+#include <iostream>
+#include <sstream>
+#include <string>
+#include <cassert>
+#include <iterator>
+#include <vector>
+#include <iomanip>
+#include <complex>
+#include <algorithm>
+#include <cmath>
+#include <exception>
+
+#endif
